🚚 Rename: 아바타 사이즈 파일명 변경 (#93)
