+++ src/components/common/Skeleton/AvatarSkeleton.tsx
@@ -1,6 +1,6 @@
 import { cn } from '@/utils'
 import { Circle } from './SkeletonItem'
-import { AVATAR_SIZE } from '@/constants/AvatarSize'
+import { AVATAR_SIZE } from '@/constants/avatar-size'
 
 function AvatarSkeleton({
   size = 'md',
