+++ src/constants/avatar-size.ts
